feat: move ML stuff to `safeds.ml.classical` package
