+++ src/safeds/ml/classical/__init__.py
@@ -1,0 +1,1 @@
+"""Classes for classical machine learning."""
