+++ src/safeds/ml/classical/classification/_ada_boost.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING
 
 from sklearn.ensemble import AdaBoostClassifier as sk_AdaBoostClassifier
 
-from safeds.ml._util_sklearn import fit, predict
+from safeds.ml.classical._util_sklearn import fit, predict
 
 from ._classifier import Classifier
