+++ src/safeds/ml/classical/classification/_classifier.py
