+++ src/safeds/ml/classical/classification/_gradient_boosting_classification.py
@@ -2,29 +2,29 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from sklearn.ensemble import RandomForestRegressor as sk_RandomForestRegressor
+from sklearn.ensemble import GradientBoostingClassifier as sk_GradientBoostingClassifier
 
-from safeds.ml._util_sklearn import fit, predict
+from safeds.ml.classical._util_sklearn import fit, predict
 
-from ._regressor import Regressor
+from ._classifier import Classifier
 
 if TYPE_CHECKING:
     from safeds.data.tabular.containers import Table, TaggedTable
 
 
-class RandomForest(Regressor):
-    """Random forest regression."""
+class GradientBoosting(Classifier):
+    """Gradient boosting classification."""
 
     def __init__(self) -> None:
-        self._wrapped_regressor: sk_RandomForestRegressor | None = None
+        self._wrapped_classifier: sk_GradientBoostingClassifier | None = None
         self._feature_names: list[str] | None = None
         self._target_name: str | None = None
 
-    def fit(self, training_set: TaggedTable) -> RandomForest:
+    def fit(self, training_set: TaggedTable) -> GradientBoosting:
         """
-        Create a copy of this regressor and fit it with the given training data.
+        Create a copy of this classifier and fit it with the given training data.
 
-        This regressor is not modified.
+        This classifier is not modified.
 
         Parameters
         ----------
@@ -33,19 +33,19 @@ class RandomForest(Regressor):
 
         Returns
         -------
-        fitted_regressor : RandomForest
-            The fitted regressor.
+        fitted_classifier : GradientBoosting
+            The fitted classifier.
 
         Raises
         ------
         LearningError
             If the training data contains invalid values or if the training failed.
         """
-        wrapped_regressor = sk_RandomForestRegressor(n_jobs=-1)
-        fit(wrapped_regressor, training_set)
+        wrapped_classifier = sk_GradientBoostingClassifier()
+        fit(wrapped_classifier, training_set)
 
-        result = RandomForest()
-        result._wrapped_regressor = wrapped_regressor
+        result = GradientBoosting()
+        result._wrapped_classifier = wrapped_classifier
         result._feature_names = training_set.features.get_column_names()
         result._target_name = training_set.target.name
 
@@ -76,15 +76,15 @@ class RandomForest(Regressor):
         PredictionError
             If predicting with the given dataset failed.
         """
-        return predict(self._wrapped_regressor, dataset, self._feature_names, self._target_name)
+        return predict(self._wrapped_classifier, dataset, self._feature_names, self._target_name)
 
     def is_fitted(self) -> bool:
         """
-        Check if the regressor is fitted.
+        Check if the classifier is fitted.
 
         Returns
         -------
         is_fitted : bool
-            Whether the regressor is fitted.
+            Whether the classifier is fitted.
         """
-        return self._wrapped_regressor is not None
+        return self._wrapped_classifier is not None
