+++ src/safeds/ml/classical/regression/_gradient_boosting_regression.py
@@ -2,38 +2,29 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from sklearn.neighbors import KNeighborsClassifier as sk_KNeighborsClassifier
+from sklearn.ensemble import GradientBoostingRegressor as sk_GradientBoostingRegressor
 
-from safeds.ml._util_sklearn import fit, predict
+from safeds.ml.classical._util_sklearn import fit, predict
 
-from ._classifier import Classifier
+from ._regressor import Regressor
 
 if TYPE_CHECKING:
     from safeds.data.tabular.containers import Table, TaggedTable
 
 
-class KNearestNeighbors(Classifier):
-    """
-    K-nearest-neighbors classification.
+class GradientBoosting(Regressor):
+    """Gradient boosting regression."""
 
-    Parameters
-    ----------
-    number_of_neighbors : int
-        The number of neighbors to be interpolated with. Has to be less than or equal to the sample size.
-    """
-
-    def __init__(self, number_of_neighbors: int) -> None:
-        self._number_of_neighbors = number_of_neighbors
-
-        self._wrapped_classifier: sk_KNeighborsClassifier | None = None
+    def __init__(self) -> None:
+        self._wrapped_regressor: sk_GradientBoostingRegressor | None = None
         self._feature_names: list[str] | None = None
         self._target_name: str | None = None
 
-    def fit(self, training_set: TaggedTable) -> KNearestNeighbors:
+    def fit(self, training_set: TaggedTable) -> GradientBoosting:
         """
-        Create a copy of this classifier and fit it with the given training data.
+        Create a copy of this regressor and fit it with the given training data.
 
-        This classifier is not modified.
+        This regressor is not modified.
 
         Parameters
         ----------
@@ -42,19 +33,19 @@ class KNearestNeighbors(Classifier):
 
         Returns
         -------
-        fitted_classifier : KNearestNeighbors
-            The fitted classifier.
+        fitted_regressor : GradientBoosting
+            The fitted regressor.
 
         Raises
         ------
         LearningError
             If the training data contains invalid values or if the training failed.
         """
-        wrapped_classifier = sk_KNeighborsClassifier(self._number_of_neighbors, n_jobs=-1)
-        fit(wrapped_classifier, training_set)
+        wrapped_regressor = sk_GradientBoostingRegressor()
+        fit(wrapped_regressor, training_set)
 
-        result = KNearestNeighbors(self._number_of_neighbors)
-        result._wrapped_classifier = wrapped_classifier
+        result = GradientBoosting()
+        result._wrapped_regressor = wrapped_regressor
         result._feature_names = training_set.features.get_column_names()
         result._target_name = training_set.target.name
 
@@ -85,15 +76,15 @@ class KNearestNeighbors(Classifier):
         PredictionError
             If predicting with the given dataset failed.
         """
-        return predict(self._wrapped_classifier, dataset, self._feature_names, self._target_name)
+        return predict(self._wrapped_regressor, dataset, self._feature_names, self._target_name)
 
     def is_fitted(self) -> bool:
         """
-        Check if the classifier is fitted.
+        Check if the regressor is fitted.
 
         Returns
         -------
         is_fitted : bool
-            Whether the classifier is fitted.
+            Whether the regressor is fitted.
         """
-        return self._wrapped_classifier is not None
+        return self._wrapped_regressor is not None
