+++ src/safeds/ml/classical/regression/_regressor.py
